requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
jinja2==3.1.5
httpx==0.28.1
python-multipart==0.0.20
pydantic==2.10.6
groq==0.9.0
python-dotenv==1.0.0 